Add NewBuilder.
